+++ CHANGELOG.md
@@ -9,11 +9,20 @@ The changelog for `PresenterKit`. Also see the [releases](https://github.com/jes
 
 This release closes the [5.0.0 milestone](https://github.com/jessesquires/PresenterKit/milestone/6?closed=1).
 
+### Breaking
+
 - Swift 4.1
 - iOS 10 minimum deployment target (Dropped iOS 9)
 - Xcode 9.4
 - Upgrade Swiftlint to 0.26.0
+
+### Fixed
+
 - Respect transition context `.isCancelled` when pushing on a navigation stack (#38)
+
+### New
+
+- Add new `pop()` method extension on `UINavigationController` that accepts a `completion` parameter. Similar to existing `push()` method. (#40)
 
 4.0.0
 -----
@@ -28,7 +37,6 @@ This release closes the [4.0.0 milestone](https://github.com/jessesquires/Presen
 
 - `public struct PopoverConfig` was changed to accommodate custom frame so it can be used as anchor for the popover.  ([#27](https://github.com/jessesquires/PresenterKit/pull/27), [#26](https://github.com/jessesquires/PresenterKit/issues/26), [@psartzetakis](https://github.com/psartzetakis))
   - `PopoverConfig.Source.view(_)` was changed to `PopoverConfig.Source.view(container: frame:)`
-
 
 - `public struct DismissButtonConfig` was changed to accommodate custom images in bar button items. ([#24](https://github.com/jessesquires/PresenterKit/pull/24), [#22](https://github.com/jessesquires/PresenterKit/issues/22), [@psartzetakis](https://github.com/psartzetakis))
     - `DismissButtonConfig.text` was renamed to `DismissButtonConfig.content`